+++ src/Microsoft.Identity.Web.Certificate/SignedAssertionFromManagedIdentityCredentialLoader.cs
@@ -30,10 +30,10 @@ namespace Microsoft.Identity.Web
                     // Given that managed identity can be not available locally, we need to try to get a
                     // signed assertion, and if it fails, move to the next credentials
                     _= await managedIdentityClientAssertion!.GetSignedAssertion(CancellationToken.None);
+                    credentialDescription.CachedValue = managedIdentityClientAssertion;
                 }
                 catch (AuthenticationFailedException)
                 {
-                    credentialDescription.CachedValue = managedIdentityClientAssertion;
                     credentialDescription.Skip = true;
                     throw;
                 }