+++ changelog.md
@@ -1,3 +1,38 @@
+2.12.4
+==========
+- fix for CVE-2023-29331 in `System.Security.Cryptography.Pkcs`
+
+2.12.2
+==========
+### New Feature:
+- **Id Web now supports the MS Graph v5 SDK,** see issue [#2097](https://github.com/AzureAD/microsoft-identity-web/issues/2097) for details.
+
+2.11.1
+==========
+- Update to MSAL 4.54.1
+
+### Bug Fix:
+- **Fix bug with signed assertion for AKS**, see issue [#2252](https://github.com/AzureAD/microsoft-identity-web/pull/2252) for details.
+
+2.11.0
+==========
+- Update to MSAL 4.54.0
+
+### New Features
+- **Id Web now supports [trimming](https://learn.microsoft.com/en-us/dotnet/core/deploying/trimming/trim-self-contained)**. See [#2210](https://github.com/AzureAD/microsoft-identity-web/pull/2210)
+
+2.10.0
+==========
+- Update to Wilson 6.30.0
+
+### New features:
+- **Microsoft.Identity.Web now provides more logging in DownstreamAPI**, see [#2148](https://github.com/AzureAD/microsoft-identity-web/issues/2148) for details.
+- **OBO support for composite tokens** based assertion and sub_assertion extra query parameters. See issue [#2222](https://github.com/AzureAD/microsoft-identity-web/issues/2222) for details.
+
+### Bug fixes:
+- **Fix a regex issue** relating to same site, see [#1811](https://github.com/AzureAD/microsoft-identity-web/issues/1811) for details.
+- Bug fixes for CIAM support, see [#2218](https://github.com/AzureAD/microsoft-identity-web/pull/2218) for details.
+
 2.9.0
 ==========
 - Update to Wilson 6.29.0 and MSAL.NET 4.53.0